Remove `org_id` from `PurchaseOrderPayload`
This change removes the `org_id` field from the `PurchaseOrderPayload`,
as the organizations involved in a purchase order are defined by the
`buyer_org` andd `seller_org` fields. The organization ID of the
organization submitting the request is not necessary, as we have the
agent's public key and are able to validate as necessary using the
signer.

Signed-off-by: Shannyn Telander <telander@bitwise.io>